XGBoost imputer: Multiple imputation through XGBoost

Set up an XGBoost imputer object with specified hyperparameters and obtain multiply imputed datasets

Public fields

data

A data.frame or data.table with missing values.

xgb.params

A list of XGBoost hyperparameters.

nrounds

The maximum number of boosting iterations for XGBoost.

early_stopping_rounds

An integer value k. XGBoost training will stop if the validation performance hasn't improved for k rounds.

print_every_n

Print XGBoost evaluation information at every nth iteration if verbose > 0.

verbose

Verbose setting for XGBoost training.

Methods

Public methods

Mixgb$new()

Mixgb$impute()

Method new()

Create a new Mixgb object. This is used to set up an XGBoost imputer with chosen hyperparameters and verbose options

Usage

Mixgb$new(
  data,
  xgb.params = list(max_depth = 6, gamma = 0.1, eta = 0.3, min_child_weight = 1,
    subsample = 1, colsample_bytree = 1, colsample_bylevel = 1, colsample_bynode = 1,
    nthread = 4, tree_method = "auto", gpu_id = 0, predictor = "auto"),
  nrounds = 50,
  early_stopping_rounds = 10,
  print_every_n = 10L,
  verbose = 0
)

Arguments

data

A data.frame or data.table with missing values

xgb.params

A list of XGBoost hyperparameters.

nrounds

The maximum number of boosting iterations for XGBoost. Default: 50

early_stopping_rounds

An integer value k. XGBoost training will stop if the validation performance hasn't improved for k rounds. Default: 10.

print_every_n

Print XGBoost evaluation information at every nth iteration if verbose > 0.

verbose

Verbose setting for XGBoost training: 0 (silent), 1 (print information) and 2 (print additional information). Default: 0

Returns

An XGBoost imputer object with chosen hyperparameters and verbose options

Method impute()

Impute missing values with the pre-set XGBoost imputer and obtain multiply imputed datasets, with an option to save models for imputing new data later on.

Usage

Mixgb$impute(
  m = 5,
  maxit = 1,
  bootstrap = TRUE,
  pmm.type = "auto",
  pmm.k = 5,
  pmm.link = "prob",
  ordinalAsInteger = TRUE,
  initial.num = "normal",
  initial.int = "mode",
  initial.fac = "mode",
  save.models = FALSE,
  save.vars = NULL,
  ...
)

Arguments

m

The number of imputed datasets. Default: 5

maxit

The number of imputation iterations. Default: 1

bootstrap

Whether to use bootstrapping for multiple imputation. By default, bootstrap = TRUE. If FALSE, users are recommended to specify sampling-related hyperparameters of XGBoost to obtain imputations with adequate variability.

pmm.type

The types of predictive mean matching (PMM). Possible values:

NULL: Imputations without PMM;

0: Imputations with PMM type 0;

1: Imputations with PMM type 1;

2: Imputations with PMM type 2;

"auto" (Default): Imputations with PMM type 2 for numeric/integer variables; imputations without PMM for categorical variables.

pmm.k

The number of donors for predictive mean matching. Default: 5

pmm.link

The link for predictive mean matching binary variables

"prob" (Default): use probabilities;

"logit": use logit values.

ordinalAsInteger

Whether to convert ordinal factors to integers. The default setting ordinalAsInteger = TRUE can speed up the imputation process.

initial.num

Initial imputation method for numeric type data:

"normal" (Default);

"mean";

"median";

"mode";

"sample".

initial.int

Initial imputation method for integer type data:

"mode" (Default);

"sample".

initial.fac

Initial imputation method for factor type data:

"mode" (Default);

"sample".

save.models

Whether to save models for imputing new data later on. Default: FALSE

save.vars

Response models for variables specified in save.vars will be saved for imputing new data. Can be a vector of names or indices. By default, save.vars = NULL, response models for variables with missing values will be saved. To save all models, please specify save.vars = colnames(data).

...

Extra arguments to pass to XGBoost

Returns

If save.models = FALSE, will return a list of m imputed datasets. If save.models = TRUE, will return an object with imputed datasets, saved models and parameters.

Examples

if (FALSE) {
#use default
MIXGB <- Mixgb$new(data = nhanes3_newborn)
#obtain m multiply datasets without saving models
imputed.data <- MIXGB$impute(m = 5)

#obtain m multiply imputed datasets and save models for imputing new data later on
imputed.obj <- MIXGB$impute(m = 5, save.models = TRUE)
}