Impute new data with a saved mixgb imputer object

Impute new data with a saved mixgb imputer object

Usage

impute_new(object, newdata, initial.newdata = FALSE, pmm.k = NULL, m = NULL)

Arguments

object

A saved imputer object created by mixgb(..., save.models = TRUE)

newdata

A data.frame or data.table. New data with missing values.

initial.newdata

Whether to use the information of the new data to initially impute new data. By default, this is set to FALSE, the original data passed to MIXGB$new() will be used for initial imputation.

pmm.k

The number of donors for predictive mean matching. If NULL (the default), the pmm.k value in the saved imputer object will be used.

m

The number of imputed datasets. If NULL (the default), the m value in the saved imputer object will be used.

Value

A list of `m` imputed datasets for new data.

Examples

# \donttest{
set.seed(2022)
n <- nrow(nhanes3_newborn)
idx <- sample(1:n, size = round(0.7 * n), replace = FALSE)
train.data <- nhanes3_newborn[idx, ]
test.data <- nhanes3_newborn[-idx, ]

mixgb.obj <- mixgb(data = train.data, m = 2, save.models = TRUE)
#> mixgb with bootstrap: saving models and imputing set -- 1 -- 2
#obtain m imputed datasets for train.data
train.imputed <- mixgb.obj$imputed.data
train.imputed
#> [[1]]
#>       HSHSIZER HSAGEIR HSSEX DMARACER DMAETHNR DMARETHN BMPHEAD BMPRECUM BMPSB1
#>    1:        7       2     1        1        1        3    44.6     68.7    6.6
#>    2:        4       3     2        2        3        2    42.6     67.1    8.8
#>    3:        3       9     2        2        3        2    46.5     64.3    8.6
#>    4:        3       9     2        1        3        1    46.2     68.5   10.8
#>    5:        5       4     1        1        3        1    44.7     63.0    6.0
#>   ---                                                                          
#> 1471:        3       9     2        1        3        1    45.0     69.7    8.1
#> 1472:        4       3     2        1        3        1    41.0     63.5   11.4
#> 1473:        5       4     2        1        3        1    42.6     69.7    7.0
#> 1474:        4       8     1        1        1        3    44.2     72.9    9.3
#> 1475:        6       5     1        2        3        2    43.8     65.3    8.8
#>       BMPSB2 BMPTR1 BMPTR2 BMPWT DMPPIR HFF1 HYD1
#>    1:    6.8    9.2    9.4  9.85  1.701    2    1
#>    2:    8.8   13.3   12.2  8.70  0.102    2    1
#>    3:    8.0   10.4    9.2  8.00  0.359    1    3
#>    4:   10.0   16.6   16.0  8.98  0.561    1    3
#>    5:    5.8    9.0    9.0  7.60  2.379    2    1
#>   ---                                            
#> 1471:    9.4   11.4   11.4  9.15  6.770    2    1
#> 1472:   10.4   11.8    9.8  7.40  2.267    2    1
#> 1473:    6.6    6.8    6.4  7.75  2.587    1    1
#> 1474:    9.0   11.1   12.4  9.65  1.050    2    2
#> 1475:   11.4   11.0   10.0  8.65  0.175    1    2
#> 
#> [[2]]
#>       HSHSIZER HSAGEIR HSSEX DMARACER DMAETHNR DMARETHN BMPHEAD BMPRECUM BMPSB1
#>    1:        7       2     1        1        1        3    43.7     68.5    6.8
#>    2:        4       3     2        2        3        2    42.6     67.1    8.8
#>    3:        3       9     2        2        3        2    46.5     64.3    8.6
#>    4:        3       9     2        1        3        1    46.2     68.5   10.8
#>    5:        5       4     1        1        3        1    44.7     63.0    6.0
#>   ---                                                                          
#> 1471:        3       9     2        1        3        1    45.0     69.7    8.1
#> 1472:        4       3     2        1        3        1    41.0     63.5   11.4
#> 1473:        5       4     2        1        3        1    42.6     69.7    7.0
#> 1474:        4       8     1        1        1        3    44.2     72.9    9.3
#> 1475:        6       5     1        2        3        2    43.8     65.3    8.8
#>       BMPSB2 BMPTR1 BMPTR2 BMPWT DMPPIR HFF1 HYD1
#>    1:    6.8    8.6    9.2  9.50  1.701    2    1
#>    2:    8.8   13.3   12.2  8.70  0.102    2    1
#>    3:    8.0   10.4    9.2  8.00  0.359    1    3
#>    4:   10.0   16.6   16.0  8.98  0.561    1    3
#>    5:    5.8    9.0    9.0  7.60  2.379    2    1
#>   ---                                            
#> 1471:    9.4   11.4   11.4  9.15  6.770    2    1
#> 1472:   10.4   11.8    9.8  7.40  2.267    2    1
#> 1473:    6.6    6.8    6.4  7.75  2.587    1    1
#> 1474:    9.0   11.1   12.4  9.65  1.050    2    2
#> 1475:   11.4   11.0   10.0  8.65  0.175    1    2
#> 

#use the saved imputer to impute new data
test.imputed <- impute_new(object = mixgb.obj, newdata = test.data)
#> Imputing new data with mixgb:  set -- 1 -- 2

# }