Show the imputed values of missing entries in a variable

Show m sets of imputed values for a specified variable.

Usage

imputed_var(imputation.list, var.name, original.data, true.values = NULL)

Arguments

imputation.list

A list of m imputed datasets returned by the mixgb imputer

var.name

The name of a variable of interest

original.data

The original data with missing data

true.values

A vector of the true values (if known) of the missing values. In general, this is unknown.

Value

A data.table with m columns, each column represents the imputed values of all missing entries in the specified variable. If true.values is provided, the last column will be the true values of the missing values.

Examples

MIXGB <- Mixgb$new(data = nhanes3_newborn)
imputed.data <- MIXGB$impute(m = 5)
#> mixgb with bootstrap: imputing set -- 1 -- 2 -- 3 -- 4 -- 5
imputed.BMPHEAD <- imputed_var(imputation.list = imputed.data, var.name = "BMPHEAD",
                               original.data = nhanes3_newborn)
imputed.BMPHEAD
#>        m1   m2   m3   m4   m5
#>   1: 43.2 42.9 44.8 43.2 41.6
#>   2: 42.0 40.8 42.4 42.0 42.1
#>   3: 42.9 42.8 43.2 42.2 42.8
#>   4: 44.3 43.8 45.7 43.9 43.1
#>   5: 46.6 44.7 45.3 46.0 46.4
#>  ---                         
#> 120: 46.0 45.7 45.2 45.9 44.2
#> 121: 44.8 44.4 43.8 44.4 44.6
#> 122: 40.9 41.6 41.3 41.9 43.8
#> 123: 43.8 44.0 42.8 44.7 44.8
#> 124: 46.7 45.5 43.4 45.2 45.0