Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds. This is a preliminary feature and functions to pre-tune other hyperparameters are under development.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_covariates = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_covariates

The names or indices of selected covariates. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds and evaluation.log

Examples

cv.results <- mixgb_cv(data = nhanes3_newborn)
#> [1]	train-rmse:5.619881+0.024757	test-rmse:5.613416+0.120455 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:3.979239+0.018594	test-rmse:3.982946+0.101815 
#> [3]	train-rmse:2.835465+0.015154	test-rmse:2.849681+0.094977 
#> [4]	train-rmse:2.040137+0.013184	test-rmse:2.065831+0.087712 
#> [5]	train-rmse:1.490774+0.012399	test-rmse:1.528652+0.085612 
#> [6]	train-rmse:1.113456+0.011182	test-rmse:1.177239+0.084232 
#> [7]	train-rmse:0.859092+0.010545	test-rmse:0.955061+0.083804 
#> [8]	train-rmse:0.689090+0.010129	test-rmse:0.819175+0.084195 
#> [9]	train-rmse:0.578671+0.008773	test-rmse:0.742699+0.085757 
#> [10]	train-rmse:0.508191+0.006988	test-rmse:0.700854+0.087197 
#> [11]	train-rmse:0.460863+0.005250	test-rmse:0.680771+0.085942 
#> [12]	train-rmse:0.434085+0.006674	test-rmse:0.669201+0.086098 
#> [13]	train-rmse:0.415328+0.007855	test-rmse:0.664217+0.085488 
#> [14]	train-rmse:0.400885+0.008993	test-rmse:0.662343+0.086229 
#> [15]	train-rmse:0.391729+0.008591	test-rmse:0.661021+0.085913 
#> [16]	train-rmse:0.382330+0.009204	test-rmse:0.660044+0.084307 
#> [17]	train-rmse:0.370998+0.007211	test-rmse:0.661285+0.084390 
#> [18]	train-rmse:0.366578+0.006106	test-rmse:0.660988+0.084559 
#> [19]	train-rmse:0.360615+0.007860	test-rmse:0.662062+0.084780 
#> [20]	train-rmse:0.349350+0.004944	test-rmse:0.661857+0.084711 
#> [21]	train-rmse:0.343379+0.006598	test-rmse:0.662397+0.084318 
#> [22]	train-rmse:0.336884+0.005016	test-rmse:0.662553+0.083950 
#> [23]	train-rmse:0.331603+0.004763	test-rmse:0.662599+0.084383 
#> [24]	train-rmse:0.325129+0.005286	test-rmse:0.663477+0.083746 
#> [25]	train-rmse:0.321539+0.004247	test-rmse:0.663921+0.084880 
#> [26]	train-rmse:0.317677+0.002526	test-rmse:0.664360+0.085070 
#> Stopping. Best iteration:
#> [16]	train-rmse:0.382330+0.009204	test-rmse:0.660044+0.084307
#> 
cv.results$best.nrounds
#> [1] 16

MIXGB <- Mixgb$new(data = nhanes3_newborn, nrounds = cv.results$best.nrounds)
imputed.data <- MIXGB$impute(m = 5)
#> mixgb with bootstrap: imputing set -- 1 -- 2 -- 3 -- 4 -- 5