Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or indices of selected features. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

cv.results <- mixgb_cv(data = nhanes3)
#> [1]	train-rmse:48.032285+0.052644	test-rmse:48.009961+0.235292 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:33.790307+0.039069	test-rmse:33.769299+0.269075 
#> [3]	train-rmse:23.812667+0.029296	test-rmse:23.798414+0.282772 
#> [4]	train-rmse:16.832128+0.026989	test-rmse:16.820986+0.327570 
#> [5]	train-rmse:11.968092+0.030314	test-rmse:11.976539+0.326834 
#> [6]	train-rmse:8.593018+0.040537	test-rmse:8.601800+0.348012 
#> [7]	train-rmse:6.270329+0.052619	test-rmse:6.329023+0.361073 
#> [8]	train-rmse:4.684832+0.058722	test-rmse:4.803712+0.406815 
#> [9]	train-rmse:3.626624+0.069196	test-rmse:3.852995+0.442071 
#> [10]	train-rmse:2.905563+0.082085	test-rmse:3.272019+0.488166 
#> [11]	train-rmse:2.396417+0.078409	test-rmse:2.948746+0.506970 
#> [12]	train-rmse:2.055029+0.076639	test-rmse:2.751801+0.522064 
#> [13]	train-rmse:1.820365+0.067589	test-rmse:2.659842+0.534106 
#> [14]	train-rmse:1.646174+0.057933	test-rmse:2.609659+0.532920 
#> [15]	train-rmse:1.510708+0.059628	test-rmse:2.578243+0.554851 
#> [16]	train-rmse:1.396963+0.067337	test-rmse:2.560365+0.558715 
#> [17]	train-rmse:1.302385+0.084581	test-rmse:2.555470+0.545869 
#> [18]	train-rmse:1.226626+0.079254	test-rmse:2.565982+0.540956 
#> [19]	train-rmse:1.155997+0.077813	test-rmse:2.576105+0.541660 
#> [20]	train-rmse:1.094470+0.078336	test-rmse:2.574121+0.526229 
#> [21]	train-rmse:1.053330+0.077302	test-rmse:2.577190+0.527320 
#> [22]	train-rmse:1.015246+0.083013	test-rmse:2.584709+0.517195 
#> [23]	train-rmse:0.981667+0.073410	test-rmse:2.593943+0.522199 
#> [24]	train-rmse:0.935828+0.076756	test-rmse:2.592228+0.518498 
#> [25]	train-rmse:0.905407+0.075484	test-rmse:2.593278+0.515265 
#> [26]	train-rmse:0.872310+0.077231	test-rmse:2.597805+0.512957 
#> [27]	train-rmse:0.851707+0.078039	test-rmse:2.599420+0.514450 
#> Stopping. Best iteration:
#> [17]	train-rmse:1.302385+0.084581	test-rmse:2.555470+0.545869
#> 
cv.results$best.nrounds
#> [1] 17

imputed.data <- mixgb(data = nhanes3, m = 5, nrounds = cv.results$best.nrounds)