Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or indices of selected features. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

cv.results <- mixgb_cv(data = nhanes3)
#> [1]	train-rmse:5.668994+0.029373	test-rmse:5.674090+0.129285 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:4.037627+0.028452	test-rmse:4.049914+0.147309 
#> [3]	train-rmse:2.905190+0.030719	test-rmse:2.918298+0.165520 
#> [4]	train-rmse:2.127553+0.035300	test-rmse:2.144819+0.191461 
#> [5]	train-rmse:1.599436+0.040129	test-rmse:1.632963+0.214442 
#> [6]	train-rmse:1.237208+0.040840	test-rmse:1.318104+0.223840 
#> [7]	train-rmse:0.991165+0.041249	test-rmse:1.121004+0.246966 
#> [8]	train-rmse:0.823648+0.040392	test-rmse:1.016617+0.248274 
#> [9]	train-rmse:0.709410+0.038613	test-rmse:0.956697+0.250091 
#> [10]	train-rmse:0.625397+0.036883	test-rmse:0.928860+0.247967 
#> [11]	train-rmse:0.566013+0.030967	test-rmse:0.920451+0.248473 
#> [12]	train-rmse:0.525176+0.023145	test-rmse:0.919590+0.248075 
#> [13]	train-rmse:0.498547+0.023641	test-rmse:0.925714+0.247203 
#> [14]	train-rmse:0.476845+0.019590	test-rmse:0.932537+0.247351 
#> [15]	train-rmse:0.463126+0.018952	test-rmse:0.932030+0.245844 
#> [16]	train-rmse:0.446635+0.013423	test-rmse:0.935238+0.246450 
#> [17]	train-rmse:0.429979+0.015970	test-rmse:0.945742+0.243980 
#> [18]	train-rmse:0.419346+0.016564	test-rmse:0.947974+0.242512 
#> [19]	train-rmse:0.409104+0.017049	test-rmse:0.955613+0.245361 
#> [20]	train-rmse:0.393874+0.016947	test-rmse:0.959154+0.244177 
#> [21]	train-rmse:0.378493+0.015681	test-rmse:0.958922+0.244845 
#> [22]	train-rmse:0.368074+0.012854	test-rmse:0.959723+0.243819 
#> Stopping. Best iteration:
#> [12]	train-rmse:0.525176+0.023145	test-rmse:0.919590+0.248075
#> 
cv.results$best.nrounds
#> [1] 12

imputed.data <- mixgb(data = nhanes3, m = 5, nrounds = cv.results$best.nrounds)