Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_covariates = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_covariates

The names or indices of selected covariates. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

cv.results <- mixgb_cv(data = nhanes3_newborn)
#> [1]	train-rmse:5.372499+0.022807	test-rmse:5.372104+0.109729 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:3.831705+0.016993	test-rmse:3.834565+0.101483 
#> [3]	train-rmse:2.771187+0.013259	test-rmse:2.779541+0.099771 
#> [4]	train-rmse:2.047236+0.011784	test-rmse:2.079989+0.095170 
#> [5]	train-rmse:1.559151+0.010257	test-rmse:1.627489+0.087381 
#> [6]	train-rmse:1.235315+0.008751	test-rmse:1.344388+0.086889 
#> [7]	train-rmse:1.025292+0.009486	test-rmse:1.176869+0.079860 
#> [8]	train-rmse:0.889905+0.009414	test-rmse:1.079786+0.076068 
#> [9]	train-rmse:0.799291+0.009837	test-rmse:1.033928+0.072676 
#> [10]	train-rmse:0.744019+0.011477	test-rmse:1.011187+0.067803 
#> [11]	train-rmse:0.708333+0.011417	test-rmse:1.001037+0.063442 
#> [12]	train-rmse:0.681524+0.012082	test-rmse:0.993738+0.058673 
#> [13]	train-rmse:0.659405+0.012532	test-rmse:0.992768+0.057425 
#> [14]	train-rmse:0.640692+0.020389	test-rmse:0.992561+0.056141 
#> [15]	train-rmse:0.630682+0.018479	test-rmse:0.993771+0.054314 
#> [16]	train-rmse:0.617074+0.018709	test-rmse:0.996912+0.054749 
#> [17]	train-rmse:0.601780+0.022217	test-rmse:0.998410+0.054683 
#> [18]	train-rmse:0.589708+0.020644	test-rmse:1.000617+0.053774 
#> [19]	train-rmse:0.579950+0.021977	test-rmse:1.001348+0.051602 
#> [20]	train-rmse:0.568403+0.021042	test-rmse:1.002327+0.050756 
#> [21]	train-rmse:0.558608+0.021817	test-rmse:1.003424+0.051715 
#> [22]	train-rmse:0.545324+0.017317	test-rmse:1.006273+0.052251 
#> [23]	train-rmse:0.530635+0.017378	test-rmse:1.007281+0.052626 
#> [24]	train-rmse:0.520222+0.016227	test-rmse:1.007863+0.053470 
#> Stopping. Best iteration:
#> [14]	train-rmse:0.640692+0.020389	test-rmse:0.992561+0.056141
#> 
cv.results$best.nrounds
#> [1] 14

MIXGB <- Mixgb$new(data = nhanes3_newborn, nrounds = cv.results$best.nrounds)
imputed.data <- MIXGB$impute(m = 5)
#> mixgb with bootstrap: imputing set -- 1 -- 2 -- 3 -- 4 -- 5