Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or indices of selected features. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

cv.results <- mixgb_cv(data = nhanes3)
#> [1]	train-rmse:48.032154+0.065481	test-rmse:48.024610+0.348786 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:33.790001+0.049311	test-rmse:33.783965+0.271289 
#> [3]	train-rmse:23.812440+0.039502	test-rmse:23.818585+0.269661 
#> [4]	train-rmse:16.831491+0.033991	test-rmse:16.831069+0.296191 
#> [5]	train-rmse:11.968465+0.036795	test-rmse:12.003287+0.311224 
#> [6]	train-rmse:8.594116+0.039124	test-rmse:8.618390+0.328951 
#> [7]	train-rmse:6.273129+0.046426	test-rmse:6.350111+0.363095 
#> [8]	train-rmse:4.693668+0.052808	test-rmse:4.817113+0.415167 
#> [9]	train-rmse:3.631729+0.059718	test-rmse:3.864391+0.442059 
#> [10]	train-rmse:2.907722+0.067101	test-rmse:3.263851+0.472811 
#> [11]	train-rmse:2.408447+0.061674	test-rmse:2.940981+0.493163 
#> [12]	train-rmse:2.059928+0.051017	test-rmse:2.764488+0.498942 
#> [13]	train-rmse:1.819514+0.055339	test-rmse:2.649808+0.512437 
#> [14]	train-rmse:1.647332+0.060825	test-rmse:2.593148+0.511188 
#> [15]	train-rmse:1.503421+0.059729	test-rmse:2.573046+0.507528 
#> [16]	train-rmse:1.408121+0.068181	test-rmse:2.555075+0.501642 
#> [17]	train-rmse:1.339889+0.093060	test-rmse:2.534960+0.489555 
#> [18]	train-rmse:1.263155+0.100203	test-rmse:2.543698+0.478786 
#> [19]	train-rmse:1.202372+0.114912	test-rmse:2.552326+0.476774 
#> [20]	train-rmse:1.156967+0.102025	test-rmse:2.579694+0.465164 
#> [21]	train-rmse:1.112169+0.108890	test-rmse:2.598613+0.455289 
#> [22]	train-rmse:1.063977+0.110443	test-rmse:2.624825+0.436167 
#> [23]	train-rmse:1.007901+0.107868	test-rmse:2.650872+0.424665 
#> [24]	train-rmse:0.959353+0.093207	test-rmse:2.656349+0.419718 
#> [25]	train-rmse:0.917001+0.097512	test-rmse:2.666824+0.424529 
#> [26]	train-rmse:0.877916+0.088775	test-rmse:2.679210+0.417533 
#> [27]	train-rmse:0.858636+0.085610	test-rmse:2.683602+0.409816 
#> Stopping. Best iteration:
#> [17]	train-rmse:1.339889+0.093060	test-rmse:2.534960+0.489555
#> 
cv.results$best.nrounds
#> [1] 17

imputed.data <- mixgb(data = nhanes3, m = 5, nrounds = cv.results$best.nrounds)