Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to obtain the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  xgb.params = list(),
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance has not improved for k rounds.

response

The name or the column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or the indices of selected features. Default: NULL (Select all the other variables in the dataset).

xgb.params

A list of XGBoost parameters. For more details, please check XGBoost documentation on parameters.

stringsAsFactors

A logical value indicating whether all character vectors in the dataset should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to be passed to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

params <- list(max_depth = 3, subsample = 0.7, nthread = 2)
cv.results <- mixgb_cv(data = nhanes3, xgb.params = params)
#> [1]	train-rmse:48.091649+0.138285	test-rmse:48.072660+0.619633 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:33.882477+0.069892	test-rmse:33.899537+0.577323 
#> [3]	train-rmse:23.887273+0.032160	test-rmse:23.925812+0.528870 
#> [4]	train-rmse:16.941981+0.047544	test-rmse:16.992690+0.460468 
#> [5]	train-rmse:12.053279+0.052306	test-rmse:12.100502+0.369969 
#> [6]	train-rmse:8.692756+0.046940	test-rmse:8.766417+0.352753 
#> [7]	train-rmse:6.352398+0.050346	test-rmse:6.447322+0.339835 
#> [8]	train-rmse:4.760667+0.064076	test-rmse:4.882071+0.326082 
#> [9]	train-rmse:3.724715+0.048163	test-rmse:3.874689+0.372463 
#> [10]	train-rmse:3.079580+0.071336	test-rmse:3.264066+0.372140 
#> [11]	train-rmse:2.669574+0.073904	test-rmse:2.899430+0.362793 
#> [12]	train-rmse:2.419663+0.059881	test-rmse:2.710994+0.366628 
#> [13]	train-rmse:2.266848+0.068356	test-rmse:2.612370+0.384572 
#> [14]	train-rmse:2.172676+0.083003	test-rmse:2.573276+0.398675 
#> [15]	train-rmse:2.117928+0.075129	test-rmse:2.554722+0.397410 
#> [16]	train-rmse:2.053014+0.083390	test-rmse:2.532105+0.403019 
#> [17]	train-rmse:2.016981+0.082778	test-rmse:2.519929+0.396371 
#> [18]	train-rmse:1.975630+0.084090	test-rmse:2.525923+0.400775 
#> [19]	train-rmse:1.947562+0.080509	test-rmse:2.539693+0.402911 
#> [20]	train-rmse:1.920968+0.084712	test-rmse:2.540088+0.390464 
#> [21]	train-rmse:1.882288+0.087954	test-rmse:2.534260+0.398733 
#> [22]	train-rmse:1.853033+0.094566	test-rmse:2.536056+0.380154 
#> [23]	train-rmse:1.825118+0.092650	test-rmse:2.544048+0.397073 
#> [24]	train-rmse:1.817145+0.091835	test-rmse:2.552250+0.395909 
#> [25]	train-rmse:1.790913+0.097821	test-rmse:2.548783+0.393545 
#> [26]	train-rmse:1.763230+0.100216	test-rmse:2.557609+0.392671 
#> [27]	train-rmse:1.732020+0.099047	test-rmse:2.566083+0.393818 
#> Stopping. Best iteration:
#> [17]	train-rmse:2.016981+0.082778	test-rmse:2.519929+0.396371
#> 
cv.results$best.nrounds
#> [1] 17

imputed.data <- mixgb(data = nhanes3, m = 3, xgb.params = params, nrounds = cv.results$best.nrounds)
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [22:01:09] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#>