Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to find the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance hasn't improved for k rounds.

response

The name or column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or indices of selected features. Default: NULL (Select all other variables in the dataset).

stringsAsFactors

A logical value indicating whether character vectors should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to pass to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

cv.results <- mixgb_cv(data = nhanes3)
#> [1]	train-rmse:5.668298+0.041814	test-rmse:5.668713+0.203248 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:4.036719+0.035606	test-rmse:4.044492+0.185217 
#> [3]	train-rmse:2.905038+0.034161	test-rmse:2.918472+0.185427 
#> [4]	train-rmse:2.125820+0.036566	test-rmse:2.149950+0.199994 
#> [5]	train-rmse:1.596627+0.040750	test-rmse:1.637756+0.216504 
#> [6]	train-rmse:1.236418+0.040479	test-rmse:1.325203+0.234772 
#> [7]	train-rmse:0.988678+0.038849	test-rmse:1.135750+0.245913 
#> [8]	train-rmse:0.816672+0.034825	test-rmse:1.029357+0.257016 
#> [9]	train-rmse:0.699538+0.032503	test-rmse:0.968694+0.259904 
#> [10]	train-rmse:0.621073+0.031943	test-rmse:0.936905+0.261324 
#> [11]	train-rmse:0.563760+0.022934	test-rmse:0.922546+0.259399 
#> [12]	train-rmse:0.521996+0.021126	test-rmse:0.922932+0.259750 
#> [13]	train-rmse:0.494639+0.023965	test-rmse:0.926017+0.258220 
#> [14]	train-rmse:0.468161+0.014905	test-rmse:0.931005+0.253192 
#> [15]	train-rmse:0.447115+0.016632	test-rmse:0.932605+0.254567 
#> [16]	train-rmse:0.435125+0.016218	test-rmse:0.936358+0.250937 
#> [17]	train-rmse:0.422984+0.013363	test-rmse:0.940935+0.249851 
#> [18]	train-rmse:0.414266+0.014562	test-rmse:0.944101+0.247952 
#> [19]	train-rmse:0.399941+0.010635	test-rmse:0.944667+0.248246 
#> [20]	train-rmse:0.387551+0.007323	test-rmse:0.950861+0.245021 
#> [21]	train-rmse:0.379431+0.010615	test-rmse:0.954223+0.242992 
#> Stopping. Best iteration:
#> [11]	train-rmse:0.563760+0.022934	test-rmse:0.922546+0.259399
#> 
cv.results$best.nrounds
#> [1] 11

imputed.data <- mixgb(data = nhanes3, m = 5, nrounds = cv.results$best.nrounds)