Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to obtain the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  xgb.params = list(),
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance has not improved for k rounds.

response

The name or the column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or the indices of selected features. Default: NULL (Select all the other variables in the dataset).

xgb.params

A list of XGBoost parameters. For more details, please check XGBoost documentation on parameters.

stringsAsFactors

A logical value indicating whether all character vectors in the dataset should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to be passed to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

params <- list(max_depth = 3, subsample = 0.7, nthread = 2)
cv.results <- mixgb_cv(data = nhanes3, xgb.params = params)
#> [1]	train-rmse:30.545517+0.035918	test-rmse:30.544277+0.271264 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:21.483645+0.025808	test-rmse:21.480103+0.231940 
#> [3]	train-rmse:15.116835+0.020760	test-rmse:15.107923+0.225282 
#> [4]	train-rmse:10.671798+0.018323	test-rmse:10.669365+0.221007 
#> [5]	train-rmse:7.570427+0.024380	test-rmse:7.566586+0.213362 
#> [6]	train-rmse:5.393886+0.008696	test-rmse:5.401710+0.216204 
#> [7]	train-rmse:3.889877+0.016558	test-rmse:3.905897+0.203964 
#> [8]	train-rmse:2.859279+0.020708	test-rmse:2.908546+0.196896 
#> [9]	train-rmse:2.174245+0.019184	test-rmse:2.246636+0.199209 
#> [10]	train-rmse:1.728843+0.015718	test-rmse:1.832092+0.206172 
#> [11]	train-rmse:1.454143+0.016778	test-rmse:1.597323+0.197925 
#> [12]	train-rmse:1.283570+0.017158	test-rmse:1.453027+0.199082 
#> [13]	train-rmse:1.185837+0.024135	test-rmse:1.387727+0.187128 
#> [14]	train-rmse:1.126033+0.024964	test-rmse:1.341096+0.183840 
#> [15]	train-rmse:1.089111+0.032074	test-rmse:1.324841+0.179840 
#> [16]	train-rmse:1.062118+0.030643	test-rmse:1.308819+0.178305 
#> [17]	train-rmse:1.042065+0.028825	test-rmse:1.308179+0.175313 
#> [18]	train-rmse:1.027044+0.028619	test-rmse:1.313738+0.175036 
#> [19]	train-rmse:1.012320+0.028990	test-rmse:1.310390+0.170543 
#> [20]	train-rmse:1.003871+0.028776	test-rmse:1.315807+0.169610 
#> [21]	train-rmse:0.992036+0.031028	test-rmse:1.319184+0.166053 
#> [22]	train-rmse:0.984058+0.028442	test-rmse:1.318963+0.165403 
#> [23]	train-rmse:0.973277+0.028034	test-rmse:1.316435+0.166257 
#> [24]	train-rmse:0.963377+0.028617	test-rmse:1.320910+0.165242 
#> [25]	train-rmse:0.954724+0.029640	test-rmse:1.317917+0.167060 
#> [26]	train-rmse:0.945546+0.031059	test-rmse:1.319450+0.164086 
#> [27]	train-rmse:0.933054+0.028400	test-rmse:1.322332+0.166577 
#> Stopping. Best iteration:
#> [17]	train-rmse:1.042065+0.028825	test-rmse:1.308179+0.175313
#> 
cv.results$best.nrounds
#> [1] 17

imputed.data <- mixgb(data = nhanes3, m = 3, xgb.params = params, nrounds = cv.results$best.nrounds)
#> mixgb null start
#> [21:36:30] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:30] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> mixgb null end
#> mixgb null start
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> mixgb null end
#> mixgb null start
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [21:36:31] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> mixgb null end