Use cross-validation to find the optimal nrounds

Use cross-validation to find the optimal nrounds for an Mixgb imputer. Note that this method relies on the complete cases of a dataset to obtain the optimal nrounds.

Usage

mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  xgb.params = list(max_depth = 3, gamma = 0, eta = 0.3, min_child_weight = 1,
    subsample = 0.7, colsample_bytree = 1, colsample_bylevel = 1, colsample_bynode = 1,
    tree_method = "auto", gpu_id = 0, predictor = "auto"),
  stringsAsFactors = FALSE,
  verbose = TRUE,
  ...
)

Arguments

data

A data.frame or a data.table with missing values.

nfold

The number of subsamples which are randomly partitioned and of equal size. Default: 5

nrounds

The max number of iterations in XGBoost training. Default: 100

early_stopping_rounds

An integer value k. Training will stop if the validation performance has not improved for k rounds.

response

The name or the column index of a response variable. Default: NULL (Randomly select an incomplete variable).

select_features

The names or the indices of selected features. Default: NULL (Select all the other variables in the dataset).

xgb.params

A list of XGBoost parameters. For more details, please check XGBoost documentation on parameters.

stringsAsFactors

A logical value indicating whether all character vectors in the dataset should be converted to factors.

verbose

A logical value. Whether to print out cross-validation results during the process.

...

Extra arguments to be passed to XGBoost.

Value

A list of the optimal nrounds, evaluation.log and the chosen response.

Examples

params <- list(max_depth = 3, subsample = 0.7, nthread = 2)
cv.results <- mixgb_cv(data = nhanes3, xgb.params = params)
#> [1]	train-rmse:5.677905+0.040466	test-rmse:5.678879+0.175816 
#> Multiple eval metrics are present. Will use test_rmse for early stopping.
#> Will train until test_rmse hasn't improved in 10 rounds.
#> 
#> [2]	train-rmse:4.049207+0.037252	test-rmse:4.050935+0.172718 
#> [3]	train-rmse:2.917553+0.032935	test-rmse:2.919857+0.192620 
#> [4]	train-rmse:2.139108+0.042418	test-rmse:2.151114+0.199500 
#> [5]	train-rmse:1.618009+0.047262	test-rmse:1.643196+0.213510 
#> [6]	train-rmse:1.275829+0.054198	test-rmse:1.306012+0.240306 
#> [7]	train-rmse:1.066198+0.061123	test-rmse:1.107366+0.248614 
#> [8]	train-rmse:0.928778+0.062937	test-rmse:0.997180+0.251699 
#> [9]	train-rmse:0.838510+0.063291	test-rmse:0.939291+0.256222 
#> [10]	train-rmse:0.781845+0.060637	test-rmse:0.897121+0.259519 
#> [11]	train-rmse:0.747521+0.062451	test-rmse:0.879998+0.262640 
#> [12]	train-rmse:0.727125+0.061439	test-rmse:0.876901+0.260893 
#> [13]	train-rmse:0.703580+0.062875	test-rmse:0.874558+0.259398 
#> [14]	train-rmse:0.692461+0.058795	test-rmse:0.874570+0.259761 
#> [15]	train-rmse:0.685681+0.060192	test-rmse:0.875066+0.258217 
#> [16]	train-rmse:0.678826+0.060895	test-rmse:0.878739+0.258593 
#> [17]	train-rmse:0.666242+0.054307	test-rmse:0.875333+0.257336 
#> [18]	train-rmse:0.662126+0.054596	test-rmse:0.876399+0.257764 
#> [19]	train-rmse:0.651361+0.054002	test-rmse:0.887410+0.255141 
#> [20]	train-rmse:0.644648+0.055591	test-rmse:0.901270+0.256329 
#> [21]	train-rmse:0.632844+0.048679	test-rmse:0.899652+0.257690 
#> [22]	train-rmse:0.623592+0.041609	test-rmse:0.908978+0.252424 
#> [23]	train-rmse:0.620537+0.040729	test-rmse:0.916099+0.249164 
#> Stopping. Best iteration:
#> [13]	train-rmse:0.703580+0.062875	test-rmse:0.874558+0.259398
#> 
cv.results$best.nrounds
#> [1] 13

imputed.data <- mixgb(data = nhanes3, m = 3, xgb.params = params, nrounds = cv.results$best.nrounds)