Show multiply imputed values for a single variable

Show m sets of imputed values for a specified variable.

Usage

show_var(imputation.list, var.name, original.data, true.values = NULL)

Arguments

imputation.list

A list of m imputed datasets returned by the mixgb imputer.

var.name

The name of a variable of interest.

original.data

The original data with missing data.

true.values

A vector of the true values (if known) of the missing values. In general, this is unknown (true.values = NULL).

Value

A data.table with m columns, each of which represents a set of imputed values for the variable of interest. If true.values is provided, an additional column will display the true values of the missing values.

Examples

# obtain m multiply datasets
params <- list(max_depth = 3, subsample = 1, nthread = 2)
mixgb.data <- mixgb(data = nhanes3, m = 3, xgb.params = params, nrounds = 20)
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 
#> [09:57:17] WARNING: src/learner.cc:767: 
#> Parameters: { "device" } are not used.
#> 

imputed.BMPHEAD <- show_var(
  imputation.list = mixgb.data, var.name = "BMPHEAD",
  original.data = nhanes3
)
imputed.BMPHEAD
#>       m1   m2   m3
#>  1: 43.3 42.7 45.6
#>  2: 44.4 44.4 43.3
#>  3: 46.3 45.7 45.3
#>  4: 44.7 43.8 43.3
#>  5: 40.9 41.4 41.4
#>  6: 42.4 41.7 42.9
#>  7: 45.5 47.3 44.9
#>  8: 42.6 47.5 45.8
#>  9: 42.9 43.2 43.2
#> 10: 43.3 44.1 43.9
#> 11: 46.2 44.8 46.4
#> 12: 42.4 42.5 45.8
#> 13: 40.3 40.9 41.2
#> 14: 46.2 45.8 46.9
#> 15: 44.3 44.0 45.0
#> 16: 46.2 46.3 46.1
#> 17: 39.3 42.2 40.2
#> 18: 42.0 44.8 46.4
#> 19: 40.2 40.9 41.8
#> 20: 43.3 43.0 46.2
#> 21: 43.9 42.9 42.2
#> 22: 41.8 44.9 41.8
#> 23: 43.4 40.4 44.7
#> 24: 44.7 44.3 46.0
#> 25: 42.3 41.8 44.2
#> 26: 40.2 42.0 40.4
#> 27: 41.8 41.1 42.8
#> 28: 46.0 42.7 44.9
#> 29: 45.9 46.1 46.7
#> 30: 42.6 40.4 42.2
#> 31: 41.3 42.3 40.2
#> 32: 45.6 45.0 44.8
#>       m1   m2   m3